+++ src/components/Donut.vue
@@ -2,9 +2,9 @@
 <div class="cdc-container" :style="placementStyles.container">
   <div class="cdc" ref="donut" :style="donutStyles">
     <donut-sections
+      v-on="sectionListeners"
       :sections="donutSections"
-      :start-angle="startAngle"
-      @section-click="emitSectionClick">
+      :start-angle="startAngle">
     </donut-sections>
     <div class="cdc-overlay" :style="overlayStyles">
       <div class="cdc-text" :style="donutTextStyles">
@@ -25,6 +25,7 @@
 </template>
 
 <script>
+import { nativeSectionEvents } from '../utils/events';
 import defaultColors from '../utils/colors';
 import { placement, placementStyles, sectionValidator } from '../utils/misc';
 import DonutSections from './DonutSections.vue';
@@ -99,7 +100,10 @@ export default {
   },
   computed: {
     donutSections() {
-      const valueTotal = this.sections.reduce((a, c) => a + c.value, 0);
+      let valueTotal = this.sections.reduce((a, c) => a + c.value, 0);
+      if (typeof valueTotal !== 'number') return [];
+      valueTotal = Number(valueTotal.toFixed(2));
+
       if (valueTotal > this.total) {
         const err = `Sum of all the sections' values (${valueTotal}) should not exceed \`total\` (${this.total})`;
         throw new Error(err);
@@ -123,7 +127,8 @@ export default {
         const color = section.color || defaultColors[currentDefaultColorIdx++];
 
         degreeArr.forEach(degree => {
-          const consumedWithCurrent = consumedDegrees + degree;
+          // limit to 2 decimal digits to avoid floating point arithmetic issues
+          const consumedWithCurrent = Number((consumedDegrees + degree).toFixed(2));
           if (consumedWithCurrent > degreesInASection) {
             const remainingDegreesInCurrentSection = degreesInASection - consumedDegrees;
 
@@ -189,6 +194,12 @@ export default {
     donutTextStyles() {
       const { fontSize } = this;
       return { fontSize };
+    },
+    sectionListeners() {
+      return nativeSectionEvents.reduce((acc, { sectionEventName }) => ({
+        ...acc,
+        [sectionEventName]: (...args) => this.emitSectionEvent(sectionEventName, ...args)
+      }), {});
     }
   },
   methods: {
@@ -206,8 +217,8 @@ export default {
         this.fontSize = widthInPx ? `${(widthInPx * scaleDownBy).toFixed(2)}px` : '1em';
       });
     },
-    emitSectionClick(section) {
-      this.$emit('section-click', section);
+    emitSectionEvent(sectionEventName, ...args) {
+      this.$emit(sectionEventName, ...args);
     }
   },
   mounted() {