+++ src/components/site/Demo.vue
@@ -11,14 +11,7 @@
     <a href="https://github.com/dumptyd/vue-css-donut-chart">GitHub</a>
   </nav>
   <div class="container-donut">
-    <donut
-      :background="background" :foreground="foreground"
-      :size="size" :unit="unit" :thickness="thickness"
-      :hasLegend="hasLegend" :legendPlacement="legendPlacement"
-      :sections="validatedSections"
-      :total="total"
-      :start-angle="startAngle"
-      @section-click="handleSectionClick">
+    <donut v-on="listeners" v-bind="donutProps">
       <div v-html="donutHTML"></div>
     </donut>
   </div>
@@ -38,7 +31,7 @@
           </div>
           <div class="control">
             <label for="size">Size</label>
-            <input name="size" type="number" class="sm" v-model.number="size">
+            <input name="size" type="number" class="sm" min="1" v-model.number="size">
             <select v-model="unit">
               <option v-for="option in unitOptions" :key="option" :value="option">
                 {{ option }}
@@ -63,7 +56,7 @@
         <div class="control-group">
           <div class="control">
             <label for="has-legend">Has legend?</label>
-            <input name="has-legend" type="checkbox" v-model="hasLegend">
+            <input id="has-legend" type="checkbox" v-model="hasLegend">
           </div>
           <div class="control">
             <label for="placement">Legend Placement</label>
@@ -89,6 +82,20 @@
       </div>
       <!-- end donut content -->
 
+      <!-- section events -->
+      <div>
+        <h3>Section events</h3>
+        <div class="control-group">
+          <div class="control" v-for="event in events" :key="event.sectionEventName">
+            <label :for="`${event.sectionEventName}-checkbox`">{{ event.sectionEventName }}</label>
+            <input :id="`${event.sectionEventName}-checkbox`" type="checkbox" v-model="event.enabled">
+          </div>
+        </div>
+
+        <div class="note">Checked events will log to console when they're triggered.</div>
+      </div>
+      <!-- end section events -->
+
       <!-- donut sections -->
       <div class="donut-sections">
         <h3>Donut sections</h3>
@@ -100,8 +107,7 @@
             <label :for="`section-value-${idx + 1}`">Value</label>
             <input
               class="sm" name="`section-value-${idx + 1}`" :min="0" :max="section.value + remaining"
-              type="number" v-model.number="section.value"
-            >
+              type="number" v-model.number="section.value">
           </div>
           <div class="control">
             <label :for="`section-label-${idx + 1}`">Label</label>
@@ -129,15 +135,19 @@
 <script>
 import Donut from '../Donut.vue';
 import colors from '../../utils/colors';
+import { nativeSectionEvents } from '../../utils/events';
 import { version } from '../../../package.json';
 import '../../styles/normalize.css';
 import '../../styles/site.css';
 
-function getRandomIntInclusive(min, max) {
+const getRandomIntInclusive = (min, max) => {
   min = Math.ceil(min);
   max = Math.floor(max);
   return Math.floor(Math.random() * (max - min + 1)) + min;
-}
+};
+
+const getObjectWithoutReactiveKeys = obj => JSON.parse(JSON.stringify(obj));
+const toFixed = num => Number(Number(num).toFixed(2));
 
 export default {
   data() {
@@ -156,6 +166,12 @@ export default {
     }));
 
     const initialConsumed = sections.reduce((a, c) => a + c.value, 0);
+
+    const initiallyEnabledEvents = ['click'];
+    const events = nativeSectionEvents.map(event => ({
+      ...event,
+      enabled: initiallyEnabledEvents.includes(event.nativeEventName)
+    }));
 
     return {
       background: '#ffffff',
@@ -178,11 +194,31 @@ export default {
       textTypeOptions,
 
       sections,
+      events,
 
       version
     };
   },
   computed: {
+    donutProps() {
+      const {
+        background, foreground,
+        size, unit, thickness,
+        hasLegend, legendPlacement,
+        validatedSections, total,
+        startAngle
+      } = this;
+      const [computedSize, computedThickness, computedTotal, computedStartAngle] =
+        [size, thickness, total, startAngle].map(val => toFixed(val));
+      return {
+        background, foreground,
+        size: computedSize > 0 ? computedSize : 200, unit,
+        thickness: computedThickness >= 0 && computedThickness <= 100 ? computedThickness : 20,
+        hasLegend, legendPlacement,
+        sections: validatedSections, total: computedTotal > 0 ? computedTotal : 100,
+        startAngle: computedStartAngle || 0
+      };
+    },
     consumed() {
       return this.sections.reduce((a, c) => a + c.value, 0);
     },
@@ -190,8 +226,18 @@ export default {
       return this.total - this.consumed;
     },
     validatedSections() {
-      if (this.consumed > this.total) return [];
+      const { consumed, total } = this;
+      if (
+        [consumed, total].some(val => typeof val !== 'number') ||
+        toFixed(this.consumed) > toFixed(this.total)
+      ) return [];
       return this.sections;
+    },
+    listeners() {
+      return this.events.filter(event => event.enabled).reduce((acc, curr) => ({
+        ...acc,
+        [curr.sectionEventName]: (...args) => this.handleSectionEvent(curr, ...args)
+      }), {});
     }
   },
   methods: {
@@ -205,8 +251,15 @@ export default {
     removeSection(idx) {
       this.sections.splice(idx, 1);
     },
-    handleSectionClick(section) {
-      window.alert(`${section.label || 'Section'} clicked`);
+    handleSectionEvent({ sectionEventName }, section, event) {
+      const info = [
+        ['-'.repeat(10)],
+        [`"${sectionEventName}" occurred on "${section.label || 'Unnamed section'}"`],
+        ['Section object:', getObjectWithoutReactiveKeys(section)],
+        ['Native event:', event]
+      ];
+      // eslint-disable-next-line no-console
+      info.forEach(args => console.log(...args));
     }
   },
   components: { Donut }